requirements.txt
# API and server
fastapi==0.104.0
uvicorn==0.23.2
pydantic==2.4.2
python-dotenv==1.0.0

# Database
supabase==1.2.0
psycopg2-binary==2.9.9

# Vector search
chromadb==0.4.18
openai==1.3.5
numpy<2.0.0

# Utilities
python-multipart==0.0.6
requests==2.31.0
tenacity==8.2.3
httpx<0.25.0,>=0.24.0 